dom/security/test/general/test_nosniff_navigation.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1428473 Support X-Content-Type-Options: nosniff when navigating</title>
  <!-- Including SimpleTest.js so we can use waitForExplicitFinish !-->
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <style>
    iframe{
      border: 1px solid orange;
    }
  </style>

  <iframe src="file_nosniff_navigation.sjs?xml"> </iframe>
  <iframe src="file_nosniff_navigation.sjs?html"></iframe>
  <iframe src="file_nosniff_navigation.sjs?css" ></iframe>
  <iframe src="file_nosniff_navigation.sjs?json"></iframe>
  <iframe src="file_nosniff_navigation.sjs?img"></iframe>
 
  <hr>
  <iframe src="file_nosniff_navigation_mismatch.sjs?html"></iframe>
  <iframe src="file_nosniff_navigation_mismatch.sjs?xml"></iframe>
  <iframe src="file_nosniff_navigation_mismatch.sjs"></iframe>

  <iframe src="file_nosniff_navigation_garbage.sjs?xml"> </iframe>
  <iframe src="file_nosniff_navigation_garbage.sjs?html"></iframe>
  <iframe src="file_nosniff_navigation_garbage.sjs?css" ></iframe>
  <iframe src="file_nosniff_navigation_garbage.sjs?json"></iframe>
  <iframe src="file_nosniff_navigation_garbage.sjs?img"></iframe>
 

</head>
<body>

<!-- add the two script tests -->
<script id="scriptCorrectType"></script>
<script id="scriptWrongType"></script>

<script class="testbody" type="text/javascript">
/* Description of the test:
 * We're testing if Firefox respects the nosniff Header for Top-Level 
 * Navigations.
 * If Firefox cant Display the Page, it will prompt a download 
 * and the URL of the Page will be about:blank.
 * So we will try to open different content send with
 * no-mime, mismatched-mime and garbage-mime types.
 * 
 */

SimpleTest.waitForExplicitFinish();

window.addEventListener("load", ()=>{
  let iframes = Array.from(document.querySelectorAll("iframe"));

  iframes.forEach( frame => {
    let result = frame.contentWindow.document.URL == "about:blank";
    let sniffTarget = (new URL(frame.src)).search;
    ok(result, `${sniffTarget} - was not Sniffed`);
  });
  
  SimpleTest.finish();
});
</script>
</body>
</html>
